trial

+++ 0x02-functions_nested_loops/main.h
@@ -3,4 +3,4 @@ void print_alphabet(void);
 void print_alphabet_x10(void);
 int _islower(int c);
 int _isalpha(int c);
-signed int print_sign(signed int n);
+signed double int print_sign(signed double int n);
